refactor: rename onchain to nut30
