+++ crates/cashu/src/nuts/mod.rs
@@ -31,7 +31,7 @@ pub mod nut27;
 #[cfg(feature = "wallet")]
 pub mod nut28;
 pub mod nut29;
-pub mod nut_onchain;
+pub mod nut30;
 
 mod auth;
 
@@ -89,7 +89,7 @@ pub use nut27::{
     backup_filter_params, create_backup_event, decrypt_backup_event, derive_nostr_keys, MintBackup,
 };
 pub use nut29::{BatchCheckMintQuoteRequest, BatchMintRequest, Settings as Nut29Settings};
-pub use nut_onchain::{
+pub use nut30::{
     MeltOnchainRequest, MeltQuoteOnchainRequest, MeltQuoteOnchainResponse, MintQuoteOnchainRequest,
     MintQuoteOnchainResponse,
 };
